♻️(ingestion-presentation) move files related to archive offer
